permission

+++ app/Livewire/Setup/EconomicCodes.php
@@ -14,6 +14,7 @@ class EconomicCodes extends Component
 
   public function render()
   {
+    abort_if(auth()->user()->cannot('view-economic-codes'), 403);
     $this->codes = EconomicCode::orderBy('id', 'desc')->get();
     return view('livewire.setup.economic-codes')
       ->extends('layouts.skot')
@@ -22,6 +23,7 @@ class EconomicCodes extends Component
 
   public function create()
   {
+    abort_if(auth()->user()->cannot('create-economic-codes'), 403);
     $this->resetInputFields();
     $this->openModal();
   }
@@ -46,6 +48,12 @@ class EconomicCodes extends Component
 
   public function store()
   {
+    if ($this->economic_code_id) {
+        abort_if(auth()->user()->cannot('edit-economic-codes'), 403);
+    } else {
+        abort_if(auth()->user()->cannot('create-economic-codes'), 403);
+    }
+
     $this->validate([
       'name' => 'required',
       'code' => 'required|unique:economic_codes,code,' . $this->economic_code_id,
@@ -59,7 +67,7 @@ class EconomicCodes extends Component
 
     session()->flash(
       'message',
-      $this->economic_code_id ? 'Economic Code Updated Successfully.' : 'Economic Code Created Successfully.'
+      $this->economic_code_id ? __('Economic Code Updated Successfully.') : __('Economic Code Created Successfully.')
     );
 
     $this->closeModal();
@@ -68,6 +76,7 @@ class EconomicCodes extends Component
 
   public function edit($id)
   {
+    abort_if(auth()->user()->cannot('edit-economic-codes'), 403);
     $code = EconomicCode::findOrFail($id);
     $this->economic_code_id = $id;
     $this->name = $code->name;
@@ -79,15 +88,17 @@ class EconomicCodes extends Component
 
   public function delete($id)
   {
+    abort_if(auth()->user()->cannot('delete-economic-codes'), 403);
     $this->dispatch('delete-confirmation', $id);
   }
 
   public function deleteConfirmed($id)
   {
+    abort_if(auth()->user()->cannot('delete-economic-codes'), 403);
     if (is_array($id)) {
       $id = $id['id'] ?? $id[0];
     }
     EconomicCode::find($id)->delete();
-    session()->flash('message', 'Economic Code Deleted Successfully.');
+    session()->flash('message', __('Economic Code Deleted Successfully.'));
   }
 }
